+++ pyspark/test/bigdl/test_dl_classifier.py
@@ -17,7 +17,7 @@
 from bigdl.nn.layer import *
 from bigdl.nn.criterion import *
 from bigdl.util.common import *
-from bigdl.models.ml_pipeline.dl_classifier import *
+from bigdl.models.dlframe.dl_classifier import *
 from pyspark.sql.types import *
 from pyspark.context import SparkContext
 import numpy as np
