+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/example/imageclassification/ImagePredictor.scala
@@ -18,12 +18,12 @@ package com.intel.analytics.bigdl.example.imageclassification
 import java.nio.file.Paths
 
 import com.intel.analytics.bigdl.dataset.image._
+import com.intel.analytics.bigdl.dlframe.DLClassifierModel
 import com.intel.analytics.bigdl.example.imageclassification.MlUtils._
 import com.intel.analytics.bigdl.numeric.NumericFloat
 import com.intel.analytics.bigdl.utils.{Engine, LoggerFilter}
 import org.apache.log4j.{Level, Logger}
 import org.apache.spark.SparkContext
-import org.apache.spark.ml.DLClassifierModel
 import org.apache.spark.sql.SQLContext
 
 /**
