+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/optim/DLClassifierSpec.scala
@@ -16,6 +16,7 @@
 
 package com.intel.analytics.bigdl.optim
 
+import com.intel.analytics.bigdl.dlframe.{DLClassifier, DLClassifierModel}
 import com.intel.analytics.bigdl.models.lenet.LeNet5
 import com.intel.analytics.bigdl.nn._
 import com.intel.analytics.bigdl.tensor.Tensor
