+++ spark/spark-version/1.5-plus/src/main/scala/org/apache/spark/ml/DLTransformerBase.scala
@@ -23,7 +23,7 @@ import org.apache.spark.sql.DataFrame
  * Extends MlTransformer and override process to gain compatibility with
  * both spark 1.5 and spark 2.0.
  */
-private[ml] abstract class DLTransformerBase[M <: DLTransformerBase[M]]
+abstract class DLTransformerBase[M <: DLTransformerBase[M]]
   extends Model[M] {
 
   protected def internalTransform(dataFrame: DataFrame): DataFrame
