+++ spark/spark-version/2.0/src/main/scala/org/apache/spark/ml/DLEstimatorBase.scala
@@ -47,7 +47,7 @@ trait VectorCompatibility {
  * Extends MLEstimator and override process to gain compatibility with
  * both spark 1.5 and spark 2.0.
  */
-private[ml] abstract class DLEstimatorBase[Learner <: DLEstimatorBase[Learner, M],
+abstract class DLEstimatorBase[Learner <: DLEstimatorBase[Learner, M],
     M <: DLTransformerBase[M]]
   extends Estimator[M] with HasLabelCol {
